broke out the first three individual test files
